Charts clear before updating
