+++ src/components/bar-chart/bar-chart.ts
@@ -62,7 +62,7 @@ export class BarChartComponent {
   }
 
   setChart() {
-    
+    //d3.selectAll("svg > *").remove();
     this.initSvg()
     this.initAxis();
     this.drawAxis();
